requirement.txt
amqp==2.4.0
billiard==3.5.0.5
celery==4.2.1
Django==2.1.5
kombu==4.2.2.post1
python-dateutil==2.7.5
pytz==2018.9
six==1.12.0
vine==1.2.0
